requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
Django==3.1.7
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
environ==1.0
iniconfig==1.1.1
Markdown==3.3.4
packaging==20.9
Pillow==8.1.2
pluggy==0.13.1
py==1.10.0
PyJWT==2.0.1
pyparsing==2.4.7
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
gunicorn
psycopg2
